feat(wip): Buffer,Interface class見直し中

+++ broccoli-app-rp2040/src/constants.rs
@@ -19,11 +19,11 @@ pub const USB_PRODUCT_ID: [u8; 16] = *b"wipeseals devapp";
 /// USB device version as a byte array
 pub const USB_PRODUCT_DEVICE_VERSION: [u8; 4] = *b"0001";
 /// USB device number of blocks
-pub const USB_NUM_BLOCKS: u32 = 1024;
+pub const USB_NUM_BLOCKS: usize = 1024;
 /// USB device block size
-pub const USB_BLOCK_SIZE: u32 = 512;
+pub const USB_BLOCK_SIZE: usize = 512;
 /// USB device total size
-pub const USB_TOTAL_SIZE: u32 = USB_NUM_BLOCKS * USB_BLOCK_SIZE;
+pub const USB_TOTAL_SIZE: usize = USB_NUM_BLOCKS * USB_BLOCK_SIZE;
 /// LEDCTRL channel channel size
 pub const CHANNEL_USB_TO_LEDCTRL_N: usize = 1;
 /// USB Control Transfer to Bulk Transfer channel size
@@ -38,11 +38,11 @@ pub const LOGICAL_BLOCK_SIZE: usize = USB_BLOCK_SIZE as usize;
 pub const LOGICAL_BLOCK_BUFFER_N: usize = 8;
 /// NAND page size usable
 /// TODO: broccoli-coreの値を参照して決定する。これ以外のパラメータ含む
-pub const NAND_PAGE_SIZE_USABLE: u32 = 2048;
+pub const NAND_PAGE_SIZE_USABLE: usize = 2048;
 /// NAND page size metadata
-pub const NAND_PAGE_SIZE_METADATA: u32 = 128;
+pub const NAND_PAGE_SIZE_METADATA: usize = 128;
 /// NAND page size total (usable + metadata)
-pub const NAND_PAGE_SIZE_TOTAL: u32 = NAND_PAGE_SIZE_USABLE + NAND_PAGE_SIZE_METADATA;
+pub const NAND_PAGE_SIZE_TOTAL: usize = NAND_PAGE_SIZE_USABLE + NAND_PAGE_SIZE_METADATA;
 /// NAND page buffer size
 pub const NAND_PAGE_BUFFER_SIZE: usize = NAND_PAGE_SIZE_TOTAL as usize;
 /// NAND page buffer count
